[Refactor] #2224 STR_OFFSETの型エイリアスを廃止し、ushortに統一した

+++ src/effect/spells-effect-util.h
@@ -12,7 +12,7 @@ public:
     byte speed;
     short current_hp;
     short max_hp;
-    int nickname;
+    ushort nickname;
 };
 
 extern bool sukekaku;
